web/requirements.txt
flask
flask_restful
python-dotenv
psycopg2-binary
requests
json5